word-formatter/requirements.txt
browser-use>=0.1.0
playwright>=1.44.0
rich>=13.0.0
python-dotenv>=1.0.0
python-docx>=1.2.0
